backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.9
celery>=5.4.0
redis>=5.0.4
gunicorn>=22.0.0
django-cors-headers>=4.3.0